Started work on Buoyancy simulation
Got some of the basics done, but far from a working prototype

added "toy" class (the objects the user interacts with), the fluidHeight var that will change depending on area of submerged objects. Also added code on calculating fluidHeight based on area (though I'm working on unique functions on getting the area for unusual objects).