Started Friction + Site layout changes
I started work on the friction simulation. So far, I have most of the math done, but it doesn't render it yet. It only renders the slope so far.

Made a decent amount of site-wide layout changes. The how to use section is above the simulation now, and there is a table of contents next to it. I also made it so you can add units to the slider numbers.